Added a effect sound for your swordfighting.
the "jukebox" provides a timer, which decides, how much time has be be gone until the sound can be played again.
it also provides a delay, that means, the sound don´t has to be played exatcly when triggered. thats useful when we want to hear the sword, when it is half way of the animation (up to now, the sound delay is half of the animationDuration)